Extract chat related data objects to nah_chat
