+++ nah_chat/Cargo.toml
@@ -1,17 +1,11 @@
 [package]
-name = "nah"
+name = "nah_chat"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 license = "MPL-2.0"
 
 [dependencies]
 tokio = { version = "1.45.0", features = ["rt", "net"] }
-clap = { version = "4.5", features = ["derive"] }
 reqwest = { version = "0.12.15", default-features = false, features = ["rustls-tls", "http2", "charset"] }
-rustyline = "15.0.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-uuid = { version = "1.16.0", features = ["v4"] }
-nah_mcp_types = { path="../nah_mcp_types" }
-nah_chat = { path = "../nah_chat"}
-bytes = { version = "1.10" }
