+++ nah_chat/rustfmt.toml
@@ -1,0 +1,1 @@
+tab_spaces = 2
